feat(scoped-elements): add compatibility with scoped registries
